agent/requirements.txt
numpy==1.26.4
flask==3.0.0
requests==2.31.0
chromadb==0.4.22
openai>=1.50.0
PyYAML==6.0.1
schedule==1.2.1
pydantic==2.6.1
